Fix #1150: implicit lossless integer widening in delegate/overload conversion (+ binary operators) (#1156)
Apply the standard C# implicit, lossless integer-widening lattice (byte->int,
uint->long, int->long, ...) when matching a typed integer value — including a
lambda body — to an expected integer type during overload resolution, argument
and lambda-return conversion, and binary operators. Narrowing and non-widening
mixed pairs still require an explicit cast.

Scalar argument and assignment widening already worked. The gaps fixed here:

- OverloadResolution.ClassifyImplicit: add DelegateReturnNumericWidening so a
  func/arrow argument whose numeric return widens to a delegate parameter's
  return (e.g. Func<Item,uint32> -> Func<Item,int64>) is applicable. This makes
  LINQ Enumerable.Sum over a uint32 selector resolve to Sum(Func<T,long>)
  instead of GS0159 "Cannot find function Sum" (and stops the GS0158 member
  cascade). CompareConversions breaks ties on the delegate return types so the
  closest integral target (long) wins over float/double/decimal.

- Conversion.cs: IsFunctionToDelegateConvertible and the FunctionType ->
  FunctionType case now allow a numeric-widening return in addition to
  reference-assignable returns, fixing GS0155/GS0154 for lambda -> delegate.

- Binder emit: the lambda's return-widening conversion is materialized via the
  erased adapter (ConversionClassifier.BindConversion, the free-function
  argument loop, and the imported-extension-method path) so the produced
  delegate is created over a method whose return type already matches the
  target, gated to numeric-widening literals only so non-widening LINQ
  selectors (Where/Single/Select) keep their natural concrete signature.

- ExpressionBinder.Operators.BindBinaryExpression: directional implicit integer
  widening between two typed operands (e.g. uint32 + int64, uint8 + int32),
  composing with the #1144 constant-literal adaptation. int32 vs uint32 (neither
  widens) still reports GS0129. Spec updated accordingly.

Adds Issue1150 binder and emit (IL-verified + executed) tests. Updates one
Issue1144 negative test whose premise (uint8 + int32 errors) is exactly what
this change fixes, to a still-erroring non-widening pair (int32 vs uint32).

+++ test/Core.Tests/CodeAnalysis/Binding/Issue1144LiteralIntAdaptBinderTests.cs
@@ -126,10 +126,12 @@ public class Issue1144LiteralIntAdaptBinderTests
     [Fact]
     public void TwoTypedIntegerOperands_StillErrorsGS0129()
     {
-        // No literal involved: uint8 + int32 between two typed locals still
-        // requires an explicit cast (general conversion path untouched).
-        var source = Wrap(@"func F(x uint8, y int32) int32 {
-        return int32(x + y)
+        // No literal involved, and neither operand widens to the other:
+        // int32 + uint32 (neither is an implicit conversion to the other)
+        // still requires an explicit cast. Issue #1150 only adds DIRECTIONAL
+        // lossless widening (e.g. uint8 + int32), which does not apply here.
+        var source = Wrap(@"func F(x int32, y uint32) int64 {
+        return int64(x + y)
     }");
         Assert.Contains(Errors(source), d => d.Id == "GS0129");
     }
